Moved all the code from App.js to Home.js and encapsulated the app within a NavigationContainer.
